refac: factor out tags functionailty to a mixin
mdb_tags
