+++ pysts/app/mdb_update.py
@@ -1,3 +1,4 @@
+import pprint
 from flask import url_for, current_app
 from bento_meta.mdb import MDB
